flask/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
MarkupSafe==2.1.1
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
langchain-groq
groq
flask_cors